Create Assignment_5_Missing_Pieces.ck
Week 5's assignment. This includes a guitar model. Not sure about the guitar sound... if I get back to this code, I'll make it into a class and try some other UGens (Mandolin, maybe). 

What this this does do, though, is a text translation of the chord (eg, "Amaj7", "C#m7", etc.) to a six-string guitar fretting... in most cases, several variations on the specific fretting for the chord. In conjunction with a guitar tuning, you get the actual sound. In standard tuning, all of these chords are composed of notes in the Db Phrygian scale. 

The guitar model itself supports strumming, up or down, with random numbers controlling the actual strength of each string strike, and a small chance of missing a string (well, that does happen). Randomness is also applied to string-to-string timing. I have an adaptive low-pass filter, which keeps single-strings bright, but doesn't entirely destroy one's ears when multiple strings are hit. 

Enjoy!